+++ src/main/webapp/query-filters/filter/operator.tsx
